pyproject.toml
[tool.poetry]
name = "data-analyst"
version = "0.1"
description = "GCP Data Analyst Agent - Multi-Agent System for BigQuery Analytics and ML"
authors = ["Meltem Subasioglu <msubasioglu@google.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "data_analyst"}]


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
google-adk = "^1.0.0"
immutabledict = "^4.2.1"
sqlglot = "^26.10.1"
db-dtypes = "^1.4.2"
regex = "^2024.11.6"
tabulate = "^0.9.0"
google-cloud-aiplatform = { extras = ["adk", "agent-engines"], version = "^1.93.0" }
absl-py = "^2.2.2"
pydantic = "^2.11.3"
google-cloud-bigquery = "^3.34.0"
llama-index = "^0.12.41"
cloudpickle = "^3.0.0"
gradio = "^5.33.2"
google-cloud-firestore = "^2.21.0"



[tool.poetry.group.dev.dependencies]
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
    "evaluation",
], version = "^1.93.0" }
pytest = "^8.2.0"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
deepeval = "^3.1.0"


[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    #"error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    #'ignore:function ham\(\) is deprecated:DeprecationWarning',
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
